drop weapons fix + Infinite Nuke Improvement
I made a script that attempts to fix the issue where weapons dropped on death tend to clip into the ground and be impossible to pick up, the majority of the time.

Its a bit ugly because this makes it so weapons tend to float above the ground to avoid them clipping into the floor but its works.

Also I made some minor improvements to the infinite nuke script where it adds the EMP flash whent he nuke goes off, removes the Slow-mo effect, and removes the nuke vision. Keeping nukes fun, fast paced, and unpunishing for the entire lobby.